+++ fabric/src/main/java/mod/fuji/core/auxiliary/CollectionUtil.java
@@ -3,6 +3,8 @@ package mod.fuji.core.auxiliary;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import java.util.function.Predicate;
 import org.jetbrains.annotations.NotNull;
 
 public class CollectionUtil {
@@ -22,5 +24,15 @@ public class CollectionUtil {
         Collections.swap(list, index, index + 1);
     }
 
+    public static <T> Optional<Integer> findFirstIndex(@NotNull List<T> list, @NotNull Predicate<T> predicate) {
+        for (int i = 0; i < list.size(); i++) {
+            T t = list.get(i);
+            if (predicate.test(t)) {
+                return Optional.of(i);
+            }
+        }
+
+        return Optional.empty();
+    }
 
 }